Changes file extensions to .c++ and .h++
There's a good argument for using .c++ file extension for c++
source code files. The name of the language is c++.
The language is not called cpp, cc, or cxx.

Header files also contain c++ source code, but we want pairs of
header/implementation files, and they cannot both be called .c++.

Alternative 1: Sub extensions
 something.header.c++
 something.h.c++
 something_.c++

This gives correct but ugly file names.
It is important to not be ugly.
In the C community, then shortened the .header.c to just .h.
Since the .h++ file ending is not already in wide use, we should
be able to use that without causing confusion.

The c++ standard library uses no file extension on header files.
This looks good when we do #include statements in our .c++ files.
However, if there's every any doubt, a user or a script will have
to open up the file to check if it contains c++ source code,
if it lacks a file extension.
This defeats the purpose of file meta data and adds cognitive
overhead.

So. In situations where there cannot be any doubt that
a file is a c++ header file, skip the extension.
Otherwise, use h++.