Utilizes bounding planes to optimize
Creates planes for each anchorPivot and layer, for which
points below it can impossibly intersect with the that line during
that layer. This reduces the number of calls to the
intersect(Triangle, Triangle) function.

Benchmark test runs with 3DBenchy_scale16.4.binary.stl reduces execution
times from 22 min to 22 seconds, a 98.3% improvement.