Removes removeNonTriangularTriangles()
We do this because it saves ca 30% of run time.

We did a bunch of test runs exactly like this:

\# time run linc/test-models/broken/3DBenchy_scale15.7.binary.stl linc/params-example

Non-triangular triangles were actually removed only
in the softClip() function, and the 35 triangles that got removed during
the test run were all considered to have 2 corners.
That smells like one edge became shorter than 1e-4 after the clip,
so that the operator== for Vertex considered two corners to be equal.
That is a bit bad, we don't want to construct a lot of super small
triangles.
But it's not bad enough to be worth > 1% of our run time to avoid.